Refactor forgot password form to separate component
